+++ Automate/Framework/Machines/TerrainFeatures/BushMachine.cs
@@ -48,12 +48,12 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
         /// <summary>Get the machine's processing state.</summary>
         public override MachineState GetState()
         {
-            if (!this.InSeason())
-                return MachineState.Disabled;
+            if (this.Machine.tileSheetOffset.Value == 1)
+                return MachineState.Done;
 
-            return this.Machine.tileSheetOffset.Value == 1
-                ? MachineState.Done
-                : MachineState.Processing;
+            return this.InSeason()
+                ? MachineState.Processing
+                : MachineState.Disabled;
         }
 
         /// <summary>Provide input to the machine.</summary>
@@ -101,10 +101,11 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
         /// <summary>Get whether the bush is currently in-season to produce berries or tea leaves.</summary>
         private bool InSeason()
         {
-            if (this.LastSeason != Game1.currentSeason || this.LastDay != Game1.dayOfMonth)
+            string season = Game1.GetSeasonForLocation(this.Location);
+            if (this.LastSeason != season || this.LastDay != Game1.dayOfMonth)
             {
-                this.InSeasonImpl = this.Machine.inBloom(Game1.currentSeason, Game1.dayOfMonth);
-                this.LastSeason = Game1.currentSeason;
+                this.InSeasonImpl = this.Machine.inBloom(season, Game1.dayOfMonth);
+                this.LastSeason = season;
                 this.LastDay = Game1.dayOfMonth;
             }